complete feed and post sections
